+++ ChangeLog.md
@@ -1,3 +1,7 @@
+# UNRELEAESD
+
+* Implement memoized randomRs', randoms', randomRIO', randomIO'.
+
 # 0.0.1 (2015-05-14)
 
 * Initial release.